+++ src/test/java/org/folio/circulation/api/support/LoanRequestBuilder.java
@@ -3,6 +3,7 @@ package org.folio.circulation.api.support;
 import io.vertx.core.json.JsonObject;
 import org.folio.circulation.support.http.client.IndividualResource;
 import org.joda.time.DateTime;
+import org.joda.time.Period;
 import org.joda.time.format.ISODateTimeFormat;
 
 import java.util.Random;
@@ -16,11 +17,13 @@ public class LoanRequestBuilder {
   private final DateTime loanDate;
   private final String status;
   private DateTime returnDate;
+  private final String action;
+  private final DateTime dueDate;
 
   public LoanRequestBuilder() {
     this(UUID.randomUUID(), UUID.randomUUID(), UUID.randomUUID(),
-      DateTime.parse("2017-03-06T16:04:43.000+02:00",
-        ISODateTimeFormat.dateTime()), "Open", null);
+      new DateTime(2017, 03, 06, 16, 04, 43), null, "Open", null, "checkedout"
+    );
   }
 
   private LoanRequestBuilder(
@@ -28,8 +31,9 @@ public class LoanRequestBuilder {
     UUID itemId,
     UUID userId,
     DateTime loanDate,
-    String status,
-    DateTime returnDate) {
+    DateTime dueDate, String status,
+    DateTime returnDate,
+    String action) {
 
     this.id = id;
     this.itemId = itemId;
@@ -37,6 +41,8 @@ public class LoanRequestBuilder {
     this.loanDate = loanDate;
     this.status = status;
     this.returnDate = returnDate;
+    this.action = action;
+    this.dueDate = dueDate;
   }
 
   public JsonObject create() {
@@ -52,6 +58,15 @@ public class LoanRequestBuilder {
       .put("itemId", itemId.toString())
       .put("loanDate", loanDate.toString(ISODateTimeFormat.dateTime()))
       .put("status", new JsonObject().put("name", status));
+
+    if(action != null) {
+      loanRequest.put("action", action);
+    }
+
+    if(dueDate != null) {
+      loanRequest.put("dueDate",
+        dueDate.toString(ISODateTimeFormat.dateTime()));
+    }
 
     if(status == "Closed") {
       loanRequest.put("returnDate",
@@ -69,12 +84,12 @@ public class LoanRequestBuilder {
 
   public LoanRequestBuilder withLoanDate(DateTime loanDate) {
     return new LoanRequestBuilder(this.id, this.itemId, this.userId,
-      loanDate, this.status, this.returnDate);
+      loanDate, this.dueDate, this.status, this.returnDate, this.action);
   }
 
   public LoanRequestBuilder withUserId(UUID userId) {
     return new LoanRequestBuilder(this.id, this.itemId, userId,
-      this.loanDate, this.status, this.returnDate);
+      this.loanDate, this.dueDate, this.status, this.returnDate, this.action);
   }
 
   public LoanRequestBuilder withStatus(String status) {
@@ -83,29 +98,54 @@ public class LoanRequestBuilder {
       ? this.returnDate
       : this.loanDate.plusDays(1).plusHours(4);
 
+    String action = null;
+
+    switch(status) {
+      case "Open":
+        action = "checkedout";
+        break;
+      case "Closed":
+        action = "checkedin";
+        break;
+    }
+
     return new LoanRequestBuilder(this.id, this.itemId, this.userId,
-      this.loanDate, status, defaultedReturnDate);
+      this.loanDate, this.dueDate, status, defaultedReturnDate, action);
   }
 
   public LoanRequestBuilder withId(UUID id) {
     return new LoanRequestBuilder(id, this.itemId, this.userId,
-      this.loanDate, this.status, this.returnDate);
+      this.loanDate, this.dueDate, this.status, this.returnDate, this.action);
   }
 
   public LoanRequestBuilder withItemId(UUID itemId) {
     return new LoanRequestBuilder(this.id, itemId, this.userId,
-      this.loanDate, this.status, this.returnDate);
+      this.loanDate, this.dueDate, this.status, this.returnDate, this.action);
   }
 
   public LoanRequestBuilder withReturnDate(DateTime returnDate) {
     return new LoanRequestBuilder(this.id, this.itemId, this.userId,
-      this.loanDate, this.status, returnDate);
+      this.loanDate, this.dueDate, this.status, returnDate, this.action);
   }
 
   public LoanRequestBuilder withItem(IndividualResource item) {
     return new LoanRequestBuilder(this.id, item.getId(), this.userId,
-      this.loanDate, this.status, this.returnDate);
+      this.loanDate, this.dueDate, this.status, this.returnDate, this.action);
   }
 
+  public LoanRequestBuilder withDueDate(DateTime dueDate) {
+    return new LoanRequestBuilder(this.id, this.itemId, this.userId,
+      this.loanDate, dueDate, this.status, this.returnDate, this.action);
+  }
 
+  public LoanRequestBuilder dueIn(Period period) {
+    if(this.loanDate == null) {
+      throw new IllegalStateException(
+        "Cannot use period to specify due when no loan date specified");
+    }
+
+    DateTime calculatedDueDate = this.loanDate.plus(period);
+
+    return withDueDate(calculatedDueDate);
+  }
 }